Add `warn` argument and correct append code
Add `warn` as a keyword argument in `optimise.order.recursive()`, to be
passed to `optimise.order.make_graph()`.

Correct code to append *all* equations associated with an endogenous
variable to the new `reordered` variable.

Update test case for `optimise.order.recursive()`, but note that
ordering is not yet stable: sometimes the test is passed, other times it
isn't. Possible numerical issue with PageRank algorithm that requires
further investigation.