Add utility function to merge symbol tables
New function in `FSIC.utilities`: `merge_frames()`. This function is
similar to `merge_dicts()` in the sense that it overwrites object
elements according to a user-specified order of precedence. The final
index of the object is the union of the indexes of the DataFrames to be
merged.

Add corresponding test function for coverage.