ENH: Implement graph function for topological sort
Implement `topological_sort()` function (and supporting functions) in
`analysis.graph` module and make available in the sub-package API. This
function orders the equations of a system, breaking up strongly
connected components (which contain cycles) as necessary. See notes in
docstrings for details of the implementation.

Add test for expected equation order.