BUG: Correct model initialisation from a DataFrame
Correct handling in the case in which the user initialises a model by passing a
DataFrame of inputs (which provides both exogenous values and the span of the
model instance).

Previous implementation assumed that `fsic` would interpret the implied index
as a `PeriodIndex`. Removed `if` and `assert` statements so as to be agnostic
about the underlying index data type.

Added new test and data to check for this in the future (fails under the old
treatment; passes under the new).

Minor updates to tests elsewhere to reflect switch to more general treatment
and to better identify reason for test failure.

Update another test to drop use of (now-deprecated) `DataFrame.ix`.